Include “location” information for enum values
When declaring an enum value, you must now include a “location”
specifier that defines the value's position within the overall list of
values.  The possible location specifiers are `first`, `middle`, `last`,
and `only`.  We currently use this to make sure that we include the
right number of commas when creating the array of values in the runtime
descriptor.  (C99 would let us use a trailing comma, but is more
portable, and I presume that we'll need similar location-aware
expansions for other use cases later on.)